Remove junk files that cause problems on Linux
These .orig and .bak files get removed during a make clean, which happens as
part of building a source package. The deleted files then show up as changes
from the repository state, the source package is built from a dirty working
directory, and doesn't match the repo revision it was built from. These
files never should have been checked in, so it's better to get rid of them.

Change-Id: Id153fcbb9b0a165fdb3b2a53e63af1300636f8f3
Reviewed-on: http://hydra.wycliffe.ca:9160/664
Reviewed-by: Eberhard Beilharz <eb1@sil.org>
Tested-by: Jenkins <jenkins@sil.org>