Fix LocaleStrings.exe (LT-14280)
LocaleStrings didn't get ported to use the new msbuild build process and
so still accessed Bld/GlobalInclude.xml. GlobalInclude.xml got removed
recently since it's no longer used. This broke LocaleStrings.

This change fixes LocaleStrings so that it now looks in
Build/GlobalInclude.properties to get the FieldWorks version number.
It also improves the way to find the FieldWorks directory: it now looks
in HKCU as well as in the 7.0 subkey besides HKCR and the FWROOT
environment variable.

Change-Id: Ifd836cda719a11d88770908dda38d6a9322d4878