Don't remove pbuilder builds from disk
Previously we cleaned out cancelled builds left on disk. However,
since we now have the ability to do parallel builds some of those
builds might still be in progress causing this build to fail and
the other build potentially to hang. So it's better to not to delete
those obsolete builds.

Cleaning out cancelled and old builds (> 2 days) is now done by a
daily Jenkins job.

Change-Id: I06b477089ed84483676ffc9cec5b23b1ffda1657