Automatically create link to debootstrap script
When creating a pbuilder image for a foreign distro (e.g. debian
on an ubuntu system), we have to create the symlink that links the
distro to the correct script. The system takes care of creating
these links for the distros it knows about, but not for foreign
ones that were unknown at the time the system got released.

Change-Id: I4d33e1c59e37c347ab291cb8c8552afcc7ef7916