[packaging] Allow to leave packages on server when signing
This change adds an option to leave the packages on the Jenkins
server and only download the *.changes files for signing and upload
the packages afterwards from Jenkins. This makes the process
significantly faster for bigger packages. There are a number of
requirements and setup steps necessary that are documented at the
top of the `releasepackage` script.

Change-Id: Iecf190a82f44492ec48c18db7e8f720da60c4eee