Comment upload of source package
It turns out that this probably worked the entire time; the problem
that it couldn't find `*orig.tar*` file was that for mono5-sil we had
a finalized changelog entry (i.e. it was not UNRELEASED), so the
script created a new changelog entry. That caused the *orig.tar* file
to not be included in `*source.changes` file.

Explicitly uploading the source package caused new problems because
both the source package and the binary packages (that include the
source .dsc file) reference the `*orig.tar*` file. The first
package that gets processed in the incoming directory will delete
the `*orig.tar*` file so that the other package can't find it
anymore.

Therefore this change reverts uploading the `*source.changes`
file in the hope that this eventually solves the problem.

Change-Id: Ic2254174c3957c9f216a7f4f2e8bf30b0742c30b