Add --repack and --keep options
This change adds a `--repack` option that will repack the *.deb
files for newer Ubuntu versions to remove the zstd compression
that isn't available on pso.

It also adds a `--keep` option so that the temporary directory
that we create to download the *.deb files won't get deleted.

And finally it reformats the help output to better fit in a terminal
window that is 80 characters wide.

Change-Id: Iee9da8f31913830389f23d5f950824673903b323